emoji-predictor-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
lightning-utilities==0.14.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
PyYAML==6.0.2
setuptools==78.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
requests
